FIX: skip .gitkeep when clearing temp folder

+++ paras/scripts/data_processing/temp.py
@@ -4,7 +4,9 @@ import shutil
 def clear_temp(temp_dir):
 
     for temp_name in os.listdir(temp_dir):
-        if not temp_name == '__init__.py':
+        
+        # skip .gitkeep 
+        if not temp_name == '__init__.py' and not temp_name == '.gitkeep':
             file_location = os.path.join(temp_dir, temp_name)
             if os.path.isfile(file_location):
                 os.remove(file_location)
